refactor: spliting files, sorts
