+++ src/i18n/en.json
@@ -23,16 +23,16 @@
     "-w,--verbose": "Verbose logs",
 
 
-    "UnknownKey": "Unknown key",
+    "DifferentModelVersion": "Your file model version (%%vuser%%) is not the same as current version (%%vcurrent%%)",
     "FailedToOpenFile" : "Failed to open file",
     "InvalidJsonFile": "Invalid JSON file",
-    "JsonParsingError": "JSON parsing error",
     "InvalidTotalInput": "Invalid number of input values",
-    "UnimplementedPredictiveMode": "Unimplemented predictive mode",
-    "MissingImportFile": "Missing file to import",
-    "DifferentModelVersion": "Your file model version (%%vuser%%) is not the same as current version (%%vcurrent%%)",
-    "LayerTypeNotRecognized": "LayerType is not recognized",
+    "InvalidTotalOutput": "The expected output size provided is not valid",
+    "JsonParsingError": "JSON parsing error",
     "LayerTypeInvalidInput": "Invalid LayerType for first layer: not an InputLayer",
     "LayerTypeInvalidOutput": "Invalid LayerType for last layer: not an OutputLayer",
-    "InvalidTotalOutput": "The expected output size provided is not valid"
+    "LayerTypeNotRecognized": "LayerType is not recognized",
+    "MissingImportFile": "Missing file to import",
+    "UnimplementedPredictiveMode": "Unimplemented predictive mode",
+    "UnknownKey": "Unknown key"
 }
