+++ src/i18n/it.json
@@ -22,16 +22,16 @@
     "-v,--version": "Mostra la versione corrente",
     "-w,--verbose": "Log dettagliati",
 
-    "UnknownKey": "Chiave sconosciuta",
+    "DifferentModelVersion": "La versione del tuo file modello (%%vuser%%) non è la stessa della versione corrente (%%vcurrent%%)",
     "FailedToOpenFile": "Impossibile aprire il file",
     "InvalidJsonFile": "File JSON non valido",
-    "JsonParsingError": "Errore di parsing JSON",
     "InvalidTotalInput": "Numero di valori di input non valido",
-    "UnimplementedPredictiveMode": "Modalità predittiva non implementata",
-    "MissingImportFile": "File da importare mancante",
-    "DifferentModelVersion": "La versione del tuo file modello (%%vuser%%) non è la stessa della versione corrente (%%vcurrent%%)",
-    "LayerTypeNotRecognized": "LayerType non riconosciuto",
+    "InvalidTotalOutput": "Dimensione dell'output previsto non valida",
+    "JsonParsingError": "Errore di parsing JSON",
     "LayerTypeInvalidInput": "LayerType non valido per il primo strato: non è uno strato di input",
     "LayerTypeInvalidOutput": "LayerType non valido per l'ultimo strato: non è uno strato di output",
-    "InvalidTotalOutput": "Dimensione dell'output previsto non valida"
+    "LayerTypeNotRecognized": "LayerType non riconosciuto",
+    "MissingImportFile": "File da importare mancante",
+    "UnimplementedPredictiveMode": "Modalità predittiva non implementata",
+    "UnknownKey": "Chiave sconosciuta"
 }
