+++ src/i18n/pt.json
@@ -22,16 +22,16 @@
     "-v,--version": "Mostrar a versão atual",
     "-w,--verbose": "Registros detalhados",
 
-    "UnknownKey": "Chave desconhecida",
+    "DifferentModelVersion": "A versão do seu arquivo modelo (%%vuser%%) não é a mesma que a versão atual (%%vcurrent%%)",
     "FailedToOpenFile": "Falha ao abrir o arquivo",
     "InvalidJsonFile": "Arquivo JSON inválido",
-    "JsonParsingError": "Erro de análise JSON",
     "InvalidTotalInput": "Número inválido de valores de entrada",
-    "UnimplementedPredictiveMode": "Modo preditivo não implementado",
-    "MissingImportFile": "Arquivo faltando para importar",
-    "DifferentModelVersion": "A versão do seu arquivo modelo (%%vuser%%) não é a mesma que a versão atual (%%vcurrent%%)",
-    "LayerTypeNotRecognized": "LayerType não reconhecido",
+    "InvalidTotalOutput": "Tamanho de saída esperado inválido",
+    "JsonParsingError": "Erro de análise JSON",
     "LayerTypeInvalidInput": "LayerType inválido para a primeira camada: não é uma camada de entrada",
     "LayerTypeInvalidOutput": "LayerType inválido para a última camada: não é uma camada de saída",
-    "InvalidTotalOutput": "Tamanho de saída esperado inválido"
+    "LayerTypeNotRecognized": "LayerType não reconhecido",
+    "MissingImportFile": "Arquivo faltando para importar",
+    "UnimplementedPredictiveMode": "Modo preditivo não implementado",
+    "UnknownKey": "Chave desconhecida"
 }
